\name{obtain_from_keywords}
\alias{obtain_from_keywords}
\title{obtain the pfam id, accn, description by searching keywords}
\usage{
obtain_from_keywords(keywords="xxx") or
obtain_from_keywords("xxx")
}
\description{
obtain the pfam id, accn, description by searching keywords
}
\examples{
obtain_from_keywords(keywords="xxx") or
obtain_from_keywords("xxx")
}








