\name{obtain_full_pfampdb}
\alias{obtain_full_pfampdb}
\title{Fetch the full table, inclduing pfam id, accn, description, pdb by searching keywords}
\usage{
obtain_full_pfampdb(item="xxx") or
obtain_full_pfampdb("xxx")
}
\description{
Fetch the full table, inclduing pfam id, accn, description, pdb by searching keywords
}
\examples{
obtain_full_pfampdb(item="xxx") or
obtain_full_pfampdb("xxx")
}

